Line 234 explanation: Towns shouldn't be allowed to be RIGHT next to each other, in my opinion. Two chunks sounds right. (originally 5)
Line 242 explanation: I personally thought that 8 chunks per citizen was too few. Also, mayors will have an easier time setting up plots for purchase if it is a round number. (Originally 8)


I didn't change lines 436 and 437, but I thought that it should be noted that since I will be towny admin, I may be the right person for something like that. (suggested edit: Line 436: 'true' Line 437: 'TheRadeonDude'


Lines 597 and 598 were edited because the term ally is: Combine or unite a resource or commodity with (another) for mutual benefit. Therefore meaning, they work together for a like cause. Of course, that wouldn't necessarilly mean building or destroying one's house. Therefore, it made most sense to allow allys to use things like unprotected chests and switch. Therefore, making it easier to work with someone on a project, just not trusting them with your house. If they ever did, just make them a friend. There needs to be a differentiation between ally and outsider; they're not the same thing. (Also, these changes applied to lines 628 and 629)


Line 824 seemed slightly unfair, being that plots can be taken over with as little as three people in the nation online; especially being that the town's setting is "2"...meaning that a nation is only slightly larger than a town? It didn't add up to me...


It was really informative as to the permissions and settings in the towny config to read and slightly edit this. It feels good to have an understanding of how things will work. I hope my edits were a positive force.